+++ src/marks/frame.js
@@ -1,5 +1,5 @@
 import {create} from "d3";
-import {Mark} from "../mark.js";
+import {Mark} from "../plot.js";
 import {number} from "../options.js";
 import {applyDirectStyles, applyIndirectStyles, applyTransform, offset} from "../style.js";
